feat(item_fields): added style in field selection drop down modal
added style in field selection drop down and helper method
