+++ src/Assets/Styles/multi-select.less
@@ -104,5 +104,45 @@ label:focus:not(:hover) {
 }
 
 .dropdown-menu-field-customizer {
+  color: black;
   min-width: 250px;
+  // -webkit-box-shadow: 0 3px 5px -1px rgba(0,0,0,0.2), 0 5px 8px 0 rgba(0,0,0,0.14), 0 1px 14px 0 rgba(0,0,0,0.12);
+  // box-shadow: 0 3px 5px -1px rgba(0,0,0,0.2), 0 5px 8px 0 rgba(0,0,0,0.14), 0 1px 14px 0 rgba(0,0,0,0.12);
+  box-shadow: 3px 3px 5px 6px #ccc;
+}
+
+.checkbox-data-fields-customize {
+  font-size: 16px !important;
+}
+
+.checkbox-data-fields-customize:focus {
+  outline: 1px solid !important;
+  outline-offset: 1px;
+  background-color: rgb(208, 217, 228) !important;
+  // color: black;
+}
+.checkbox-data-fields-customize:hover {
+  outline: 0px dashed !important;
+  outline-offset: 1px;
+  background-color: rgb(208, 217, 228) !important;
+  // color: black;
+}
+
+.li-data-fields-customize {
+  padding: 5px;
+}
+
+.checkbox-fields-customize-checked {
+  font-size: 17px;
+  color: darkblue;
+}
+.checkbox-fields-customize-unchecked {
+  font-size: 17px;
+  color: darkblue;
+}
+
+.modal-backdrop {
+  background-color: rgba(100, 100, 100, 0.8);
+  width: 100vw;
+  height: 100vh;
 }
